added  review  service

+++ apps/server/src/services/admin/user-management.services.ts
@@ -3,8 +3,8 @@ import { prisma } from "@palash/db-client";
 class UserManagementServices {
     async deleteUser(userId: string) {
         await Promise.all([
-            await prisma.booking.deleteMany({where: {userId}}),
-            // await prisma.review.deleteMany({where: {userId}})
+            await prisma.booking.deleteMany({where: {user_id: userId}}),
+            await prisma.review.deleteMany({where: {user_id: userId}})
         ])
         await prisma.user.delete({
             where: {
